% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Giac.R
\name{Giac}
\alias{Giac}
\title{R6 class to access to Giac}
\description{
Creates an object allowing to execute Giac commands.
}
\examples{

## ------------------------------------------------
## Method `Giac$execute`
## ------------------------------------------------

if(!is.null(find_chrome)) {
  giac <- Giac$new()
  giac$execute("2 + 3/7")
  giac$close()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Giac-new}{\code{Giac$new()}}
\item \href{#method-Giac-execute}{\code{Giac$execute()}}
\item \href{#method-Giac-close}{\code{Giac$close()}}
\item \href{#method-Giac-clone}{\code{Giac$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Giac-new"></a>}}
\if{latex}{\out{\hypertarget{method-Giac-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Giac} instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Giac$new(chromePath = find_chrome())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chromePath}}{path to the Chrome executable (or Chromium, Brave,
etc); if \code{find_chrome()} does not work, you can set the
environment variable \code{CHROMOTE_CHROME} to the path and it will
work}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Giac} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Giac-execute"></a>}}
\if{latex}{\out{\hypertarget{method-Giac-execute}{}}}
\subsection{Method \code{execute()}}{
Execute a Giac command.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Giac$execute(command)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{the command to be executed given as a character string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of the command in a character string.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if(!is.null(find_chrome)) {
  giac <- Giac$new()
  giac$execute("2 + 3/7")
  giac$close()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Giac-close"></a>}}
\if{latex}{\out{\hypertarget{method-Giac-close}{}}}
\subsection{Method \code{close()}}{
Close a Giac session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Giac$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} or \code{FALSE}, whether the session has been closed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Giac-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Giac-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Giac$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
